Prevent deadlock from read-to-write upgrade
The RW-Lock does not allow to upgrade from a read-lock
to a write-lock. When that happens a deadlock is created.
In order to prevent this we have to check if a read-lock
is held by the current thread.
