+++ src/main/java/net/dv8tion/jda/internal/utils/cache/ReadWriteLockCache.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2015-2018 Austin Keener & Michael Ritter & Florian Spieß
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.dv8tion.jda.internal.utils.cache;
+
+import net.dv8tion.jda.internal.utils.UnlockHook;
+
+import java.util.concurrent.locks.ReentrantReadWriteLock;
+
+public abstract class ReadWriteLockCache
+{
+    protected final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
+
+    public UnlockHook writeLock()
+    {
+        if (lock.getReadHoldCount() > 0)
+            throw new IllegalStateException("Unable to acquire write-lock while holding read-lock!");
+        ReentrantReadWriteLock.WriteLock writeLock = lock.writeLock();
+        writeLock.lock();
+        return new UnlockHook(writeLock);
+    }
+
+    public UnlockHook readLock()
+    {
+        ReentrantReadWriteLock.ReadLock readLock = lock.readLock();
+        readLock.lock();
+        return new UnlockHook(readLock);
+    }
+}
